perspective_reddit_bot/requirements.txt
astroid==1.6.4
backports.functools-lru-cache==1.5
certifi==2018.4.16
chardet==3.0.4
configparser==3.5.0
enum34==1.1.6
futures==3.2.0
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
numpy==1.14.3
pandas==0.23.0
praw==5.4.0
prawcore==0.14.0
pylint==1.9.0
python-dateutil==2.7.3
pytz==2018.4
PyYAML==3.13
requests==2.19.1
singledispatch==3.4.0.3
six==1.11.0
update-checker==0.16
urllib3==1.23
wrapt==1.10.11
